+++ lib/emittance/dispatchers/synchronous.rb
@@ -12,7 +12,7 @@ module Emittance
         def _process_event(event)
           registrations_for(event).each do |registration|
             event = Emittance::Middleware.down(event)
-            registration.call event
+            registration.(event)
           end
         end